Restructured files, deleted flag.hpp, option.hpp and string_literal.hpp, this code is moved to core.hpp
