+++ examples/02_singleton/main.cpp
@@ -1,5 +1,3 @@
-#include <CLArgs/flag.hpp>
-#include <CLArgs/option.hpp>
 #include <CLArgs/singleton.hpp>
 
 #include <filesystem>
